on-premises: infiniband fullly working with this setup
Infiniband is working on TOSS
4.18.0-553.56.1.1toss.t4 based on RHEL 8.10.
For this to work, most of the issue was with
respect to network firewalls, kernel modules,
and system security. Fixes here include
needing to create unique CNI names for podman,
add a flag to ignore preflight errors (for
the old kernel) and update the flannel install
to be before 0.25.x when a check for br_netfilter
was added. This used to be part of kubeadm, and
it was removed with K8s 1.30. It is not technically
needed in the podman container (it is needed on
the physical host) but since the check is done
in the container, this will fail flannel from
starting up. For the time being, we will use
an older flannel, and I will open an issue
on the repository to ask for the ability
to disable the check.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>